requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
customtkinter==5.2.1
httpx==0.25.1
apscheduler==3.10.4